requirements.txt
# Cryptography & Security
cryptography==46.0.3

# MySQL Connector
mysql-connector-python==9.5.0

# Environment management (optional)

# Utilities

bcrypt==5.0.0